Separating quad service from tri service
