+++ sarf-web-service/src/main/java/sarfwebservice/controllers/SarfController.java
@@ -1,26 +1,27 @@
 package sarfwebservice.controllers;
 
+import org.jetbrains.annotations.NotNull;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
-import sarf.verb.Root;
 import sarfwebservice.exceptions.InvalidRootException;
 import sarfwebservice.exceptions.RootNotFoundException;
 import sarfwebservice.models.RootResult;
-import sarfwebservice.services.SarfService;
-
-import java.util.List;
+import sarfwebservice.services.SarfServiceQuad;
+import sarfwebservice.services.SarfServiceTri;
 
 @RestController
 @RequestMapping("/sarf")
 public class SarfController {
 
-    private final SarfService sarfService;
+    private final SarfServiceTri sarfServiceTri;
+    private final SarfServiceQuad sarfServiceQuad;
 
     @Autowired
-    public SarfController(SarfService sarfService) {
-        this.sarfService = sarfService;
+    public SarfController(SarfServiceTri sarfServiceTri, SarfServiceQuad sarfServiceQuad) {
+        this.sarfServiceTri = sarfServiceTri;
+        this.sarfServiceQuad = sarfServiceQuad;
     }
 
     @RequestMapping("/{rootLetters}")
@@ -28,11 +29,23 @@ public class SarfController {
         if (!isValidRoot(rootLetters)) {
             throw new InvalidRootException(String.format("ينبغي تحديد جذر مناسب للكلمة - %s.", rootLetters));
         }
-        var rootResult = sarfService.getRoots(rootLetters);
+        if(rootLetters.length() == 3) {
+            return getRootResultTri(rootLetters);
+        }
+        return getRootResultQuad(rootLetters);
+    }
+
+    @NotNull
+    private RootResult getRootResultTri(String rootLetters) throws Exception {
+        var rootResult = sarfServiceTri.getRoots(rootLetters);
         if (rootResult == null || rootResult.isEmpty()) {
             throw new RootNotFoundException(String.format("لا يوجد جذر لـ: %s", rootLetters));
         }
         return rootResult;
+    }
+
+    private RootResult getRootResultQuad(String rootLetters) {
+        return null;
     }
 
     private boolean isValidRoot(String rootLetters) {
@@ -40,7 +53,7 @@ public class SarfController {
             return false;
         }
 
-        if (!sarfService.isArabic(rootLetters)) {
+        if (!sarfServiceTri.isArabic(rootLetters)) {
             return false;
         }
