+++ sarf-web-service/src/main/java/sarfwebservice/services/SarfService.java
@@ -4,5 +4,4 @@ import sarfwebservice.models.RootResult;
 
 public interface SarfService {
     boolean isArabic(String letters);
-    RootResult getRoots(String rootLetters) throws Exception;
 }
