+++ sarf-web-service/src/main/java/sarfwebservice/services/SarfServiceQuad.java
@@ -1,0 +1,7 @@
+package sarfwebservice.services;
+
+import sarfwebservice.models.RootResult;
+
+public interface SarfServiceQuad extends SarfService {
+    RootResult getRoots(String rootLetters) throws Exception;
+}
