+++ sarf-web-service/src/main/java/sarfwebservice/services/SarfServiceTri.java
@@ -1,0 +1,7 @@
+package sarfwebservice.services;
+
+import sarfwebservice.models.RootResult;
+
+public interface SarfServiceTri extends SarfService {
+    RootResult getRoots(String rootLetters) throws Exception;
+}
